+++ lib/kitchen/directions/move_exercises_to_eoc/v1.rb
@@ -18,13 +18,6 @@ module Kitchen::Directions::MoveExercisesToEOC
         sections.each do |exercise_section|
           exercise_section.first("[data-type='title']")&.trash
 
-          exercise_section.exercises.each do |exercise|
-            exercise.pantry(name: :link_text).store(
-              "#{I18n.t(:exercise_label)} #{chapter.count_in(:book)}.#{exercise.count_in(:chapter)}",
-              label: exercise.id
-            )
-          end
-
           exercise_section.cut(to: exercise_clipboard)
         end
       end