+++ spec/directions/bake_footnotes/v1_spec.rb
@@ -26,6 +26,16 @@ RSpec.describe Kitchen::Directions::BakeFootnotes::V1 do
             <p><a href="#aside5" role="doc-noteref">[footnote]</a> Blah.</p>
             <aside id="aside5" type="footnote">Footnote content 5</aside>
           </div>
+          <div data-type="composite-chapter">
+            <div data-type="composite-page">
+              <p><a href="#aside7" role="doc-noteref">[footnote]</a> Blah.</p>
+              <aside id="aside7" type="footnote">Footnote content 7</aside>
+            </div>
+            <div data-type="composite-page">
+              <p><a href="#aside8" role="doc-noteref">[footnote]</a> Blah.</p>
+              <aside id="aside8" type="footnote">Footnote content 8</aside>
+            </div>
+          </div>
         </div>
         <div data-type="page">
           <p><a href="#aside6" role="doc-noteref">[footnote]</a> Blah.</p>
@@ -58,6 +68,16 @@ RSpec.describe Kitchen::Directions::BakeFootnotes::V1 do
           <div data-type="composite-page">
             <p><a href="#aside5" role="doc-noteref">4</a> Blah.</p>
             <aside id="aside5" type="footnote"><div data-type="footnote-number">4</div>Footnote content 5</aside>
+          </div>
+          <div data-type="composite-chapter">
+            <div data-type="composite-page">
+              <p><a href="#aside7" role="doc-noteref">5</a> Blah.</p>
+              <aside id="aside7" type="footnote"><div data-type="footnote-number">5</div>Footnote content 7</aside>
+            </div>
+            <div data-type="composite-page">
+              <p><a href="#aside8" role="doc-noteref">6</a> Blah.</p>
+              <aside id="aside8" type="footnote"><div data-type="footnote-number">6</div>Footnote content 8</aside>
+            </div>
           </div>
         </div>
         <div data-type="page">